Updated: include a canvas with 2 buttons over AR
Updates the overall workflow of the app, allowing to theoreticaly place an object, and then move it somewhere else, with 2 buttons on a 3D canvas placed over the visual in the sceneHitTest GameObject. Allowing to reiniiate he HitTest when needed, and allowing you to stopHitTest withthe click of a button.